% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleAPT-package.r
\docType{package}
\name{nimbleAPT}
\alias{nimbleAPT}
\title{nimbleAPT}
\description{
A collection of NIMBLE functions for adaptive parallel tempering.
}
\section{Adaptive parallel tempering (APT) for NIMBLE. Adapted from NIMBLE'S MCMC suite to enable APT specific features (i.e. an adaptive temperature ladder).}{
NA
}

\seealso{
\url{https://r-nimble.org/}
}
