% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APT_functions.R
\name{plot.tempTraj}
\alias{plot.tempTraj}
\title{plot.tempTraj}
\usage{
\method{plot}{tempTraj}(apt)
}
\arguments{
\item{apt}{- an APT object generated by buildAPT}
}
\value{
a plot.
}
\description{
Plot the trajectories of a temperature ladder of an adaptive parallel tempering algorithm
}
\details{
Returns two plots, one with T~iterations, the other log10(T)~iterations.
}
\author{
David Pleydell
}
