% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APT_functions.R
\name{resizeVector}
\alias{resizeVector}
\title{Resize a vector.}
\usage{
resizeVector(x, length)
}
\arguments{
\item{x}{A vector.}

\item{length}{Desired length of the new vector.}
}
\value{
A vector.
}
\description{
Returns a resized version of a vector.
}
\details{
If the new vector is longer, new elements are initialised as zeros. Otherwise, the new vector is a clipped version of the old vector.
}
\examples{
resizeVector(1:10, 11)
resizeVector(1:11, 10)

x = nimNumeric(length=10)
x = resizeVector(x, length=11)
length(x)
}
\author{
David Pleydell
}
