% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleAPT-package.r
\name{nimbleAPT}
\alias{nimbleAPT}
\title{nimbleAPT}
\description{
A collection of nimble functions for adaptive parallele tempering.
}
\section{Adaptive parallel tempering (APT) for nimble. Adapted from nimble's MCMC suite to enable APT specific features (i.e. an adaptive temperature ladder).}{
NA
}

\seealso{
\url{https://r-nimble.org/}
_PACKAGE
}
